requirements.txt
streamlit
openai
huggingface_hub
Pillow
supervision
ultralytics
openai-whisper
torch
torchaudio
torchvision